requirements.txt
asgiref==3.3.0
Django==3.1.3
django-constance==2.7.0
django-modeltranslation==0.16
django-picklefield==3.0.1
Pillow==8.0.1
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
